+++ src/metagit/core/release/__init__.py
@@ -1,0 +1,20 @@
+#!/usr/bin/env python
+"""Release and version check utilities."""
+
+from metagit.core.release.models import (
+    InstallMethod,
+    LatestReleaseInfo,
+    VersionCheckResult,
+    VersionUpgradeResult,
+)
+from metagit.core.release.release_check_service import ReleaseCheckService
+from metagit.core.release.upgrade_service import VersionUpgradeService
+
+__all__ = [
+    "InstallMethod",
+    "LatestReleaseInfo",
+    "ReleaseCheckService",
+    "VersionCheckResult",
+    "VersionUpgradeResult",
+    "VersionUpgradeService",
+]